refactor: start porting over methods
